refactor: move run service and routes to api folder (#155)
