+++ src/api/mod.rs
@@ -1,1 +1,2 @@
+pub mod app;
 pub mod routes;
