+++ tests/common/mod.rs
@@ -2,13 +2,13 @@ use std::collections::HashMap;
 
 use actix_web::{body::MessageBody, dev::ServiceFactory, web::Data, App};
 use social_service::{
+    api::app::get_app_router,
     components::{
         app::AppComponents,
         configuration::{Config, Database},
         database::{DatabaseComponent, DatabaseComponentImplementation},
         synapse::{WhoAmIResponse, WHO_AM_I_URI},
     },
-    get_app_router,
 };
 use sqlx::{Connection, Executor, PgConnection, PgPool};
 use wiremock::{
